Bell Curve done
- Bell Curve weight distribution made

+++ components/spawn_curve.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 class SpawnCurve(BaseComponent):
     parent: Actor
 
-    def __init__(self, min_prob: float = 0.0, peak_prob: float = 1.0, start_floor: int = 1, 
+    def __init__(self, min_prob: int = 0.0, peak_prob: int = 1.0, start_floor: int = 1, 
                  peak_floor: int = 5, end_floor: int = 10):
         self.min_prob = min_prob
         self.peak_prob = peak_prob
